+++ R/monitor_subsetByDistance.R
@@ -31,8 +31,9 @@ monitor_subsetByDistance <- function(ws_monitor, longitude=NULL, latitude=NULL, 
   distanceMask <- distanceVector <= radius
   
   if ( !(any(distanceMask)) ) {
-    meta <- NULL
-    data <- NULL
+    meta <- createEmptyMetaDataframe(rows=0)
+    data <- as.data.frame(ws_monitor$data$datetime)
+    names(data) <- 'datetime'
   } else {
     meta <- ws_monitor$meta[distanceMask,]
     data <- monitor_subsetData(ws_monitor$data, monitorIDs=meta$monitorID, dropMonitors=TRUE)