+++ NEWS.md
@@ -26,6 +26,8 @@ from the US EPA, AirNow, AIRSIS, WRCC and others.
  * new `monitor_join()` function allows you to merge the data of ws_monitor objects with shared monitorIDs
  * renamed `CarmelValley` dataset to `Carmel_Valley`
  * converted all uses of `GMT` to `UTC`
+ * new `createEmptyMetaDataframe()` function
+ * `monitor_subset~()` functions now return 'meta' dataframes with zero rows and 'data' dataframes with a single `datetime` column rather than `NULL` when no monitors exist in the subset.
 
 ## Version 0.99 -- Regularized and Consistent (beta)